BUG FIXED. Get an error when creating a task when all tasks have been deleted.
- Error:
When all the tasks have been deleted and you try to create a new one, it looks for the id of the last existing one, but when there is none, it jumps an error.

- Solution:
To solve it, it is previously checked if there are saved tasks, if not, the id of the new task will be 1, otherwise it will look for the id of the last task and will add 1 to it.