fix(config): the command `npm run install` does not install the packages globally
Error:
- No good update between `NVM` and `VOLTA` has been performed.
Recently I was using `NVM` as "Manage Multiple Node Versions" and I replaced it with `VOLTA`. It does exactly the same and is more modern but has some advantages like being able to include in the `package.json` file the stable versions of `NPM` and `NODE`.
But the `preinstall-global-dependencies.js` file was only prepared for use with `NVM`. This file is a script that allows to declare the global packages that must be installed for the project to work.
All these changes were made to give the user visibility of the packages and their versions that can be used at the moment.
The `@commitlint/cli`, `@commitlint/config-conventional`, `commitizen` and `cz-conventional-changelog` packages, for the moment, I had to install globally, because the `echo 'feat: create user button' | commitlint` and `git cz` commands did not work.

Solution:
- Update `preinstall-global-dependencies.js` file for install packages globally with `VOLTA`.
When `VOLTA` is being used in the `package.json` file, it displays packages already installed globally in `VOLTA`, and whether or not they are installed, it installs with `VOLTA` the packages declared in the `globalDependencies` object.
So far I have not been able to extract from `VOLTA` a list of installed packages in json format with their name as `key` and their version as `value`. Since I can't check beforehand which packages are already installed I just install them all.

References:
- https://dev.to/stefirosca/7-tips-tricks-to-make-your-console-log-output-stand-out-389g