BUGFIX. Update `prettier` configuration
Steps:
- Use the configuration of `nuxt-course` repository.
Change `proseWrap`, `quoteProps`, `semi`, `singleQuote` and `trailingComma` rules to another values.
Add missing `singleAttributePerLine` rule.
Use a matrix for override files.
Override `printWidth` rule with another value for the `json` files.

Description:
- `"proseWrap": "preserve"` was changed to `"proseWrap": "always"`:
This means that automatic line wrapping will now be applied to paragraphs, preserving the original structure of the text.
- `"semi": "false"` was changed to `"semi": "true"`:
This configuration will always add a semicolon automatically at the end of the statements.
- `"quoteProps": "as-needed"` was changed to `"quoteProps": "preserve"`:
This setting will always add quotation marks to object property names in the code.
- `"singleQuote": true` was changed to `"singleQuote": false`:
This configuration was modified to avoid conflicts with the `eslint` configuration, which required double quotes instead of single quotes.
- The `"singleAttributePerLine": true` rule was added:
This ensures that each attribute is written on a separate line, improving code readability and facilitating the identification of changes in individual attributes.
- `"trailingComma": "none"` was changed to `"trailingComma": "all"`:
Now, a comma will always be added at the end of the list of elements in arrays and objects, enhancing readability and facilitating the review of changes.
- Specific configuration was added for existing `.json` files using the `"printWidth": 0` rule:
This means that there will be no line length limits for JSON files, preventing automatic adjustments.