BUGFIX. The `json` files are not formatted with `eslint`
Error:
- The `json` files must have the properties sorted and these must be enclosed in double quotes after formatting the code.

Solution:
- Use the `nuxt-course` repository configuration.
The plugins in the list are executed in the order they appear, so this change alters the execution order of the `prettier` and `json` plugins.
When configuring the plugins, the order in which they are listed is important, as the plugins will run in that order.
The `vue` plugin will run first, followed by the `prettier` plugin, and finally the `json` plugin.

Reason:
- The change has been made to correct a problem in the formatting of `json` files when using `eslint` in the project.
In the original code, the `json` plugin was not added in the `eslint` configuration. This caused `json` files to not be formatted correctly with `prettier` rules.