ci(changelog): remove `--first-release` option from `release` command
Reason:
- The `--first-release` option is designed for projects that are implementing `standard-version` for the first time and have been previously released (i.e. they already have versions and possibly Git tags). The purpose of this option is to generate the initial changelog without altering the current version of the project, since it assumes that the project already has an established version.
When you run `--standard-version` with `--first-release`, the tool understands that it should not increment (or "bump") the version of the project, since it is generating a changelog for the current state of the project, which already has a version. This is why you see the message "✖ skip version bump on first release", indicating that the version is not being bumped as part of this initial release.
In future releases, when you don't use `--first-release`, `standard-version` will work normally, incrementing the project version, updating the changelog, and creating Git tags according to your commit messages and semantic versioning rules.