fix(globalDependencies): dependencies are not installed correctly
Error:
- 🚨 ⚠️ Error: Error: Command failed: sudo npm i -g @vue/cli@^undefined
npm ERR! code EINVALIDTAGNAME
npm ERR! Invalid tag name "^undefined" of package "@vue/cli@^undefined": Tags may not have any characters that encodeURIComponent encodes.

Reason:
- In the `preinstall-global-dependencies.js` file, somewhere in the code, it is checking if in the `package.json` file the dependencies of the `globalDependencies` object are also in `dependencies` or in `devDependencies`.
This is the case for the `@vue/cli` package, so when looking for it in `dependencies` or `devDependencies` it doesn't find it and returns `undefined` as version number.

Solution:
- Add the `isGlobalDepAlsoLocal` function to first check if the dependency exists in `dependencies` or `devDependencies`.
If the dependency does not exist, use the version number of `globalDependencies`.